Assignment WEEK I task 2 / 3
DESCRIPTION:
"In this task you will be adding appropriate formatting to the web page contents using container, row and column classes using the Bootstrap grid so that the web page is formatted to look like the figure given below.

The "About Us" title should stretch the entire width of the row.
The "Our History" part should occupy only half the width of the row for small to extra large screens, leaving space on the right side for more content to be added later. The content should be stacked for extra small screens.
The "Corporate Leadership" section should stretch the entire width of the row."